chore: rename clientoption.rs
